+++ EmptySource/Source/Runtime/Public/Fonts/Font.h
@@ -28,13 +28,13 @@ namespace ESource {
 		void Initialize(struct FileStream * File);
 
 		//* Get the index in font of given character
-		unsigned int GetGlyphIndex(const unsigned long & Character) const;
+		uint32_t GetGlyphIndex(const uint32_t & Character) const;
 
 		//* Get a copy of a Glyph
-		bool GetGlyph(FontGlyph & Glyph, const unsigned int& GlyphIndex);
+		bool GetGlyph(FontGlyph & Glyph, const uint32_t& GlyphIndex);
 
 		//* Set the glyph height for rasterization
-		void SetGlyphHeight(const unsigned int& Size) const;
+		void SetGlyphHeight(const uint32_t& Size) const;
 
 		//* Clear all the FreeType info
 		void Clear();
